fix some X64 compilation warnings

+++ src/utils/BaseUtil.cpp
@@ -34,7 +34,7 @@ uint32_t MurmurHash2(const void *key, size_t len)
     const int r = 24;
 
     /* Initialize the hash to a 'random' value */
-    uint32_t h = hash_function_seed ^ len;
+    uint32_t h = hash_function_seed ^ (uint32_t)len;
 
     /* Mix 4 bytes at a time into the hash */
     const uint8_t *data = (const uint8_t *)key;
